backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
google-adk>=1.0.0
firebase-admin>=6.5.0
pydantic>=2.9.0
python-dotenv>=1.0.0
httpx>=0.27.0
python-multipart>=0.0.12
qrcode[pil]>=8.0
Pillow>=10.4.0
websockets>=13.0
